+++ src/win32/GBMapView.cpp
@@ -500,7 +500,7 @@ LRESULT GBMapView::OnMapInfo(WPARAM wParam, LPARAM lParam)
   mapViewZoom.setColors(colors);
 
   int x = (int)(wParam & 0xffff);
-  int y = wParam >> 16;
+  int y = (int)(wParam >> 16);
   
   CString buffer;
   buffer.Format("(%d,%d)", x, y);
